refactor(observability): centralize window type augmentation
Move browser Window augmentation for third-party scripts into a single
src/types/window.d.ts and define PlausibleConfig + PlausibleFunction in
@lfx-one/shared/interfaces. PlausibleService and SegmentService no longer
declare their own globals; future integrations add one line to window.d.ts.
